refactor(segment): migrate vector-access methods to SegmentReadView
Step 3 of the SegmentReadView migration.

Move the vector-access logic onto the view in a new
`read_view/vectors.rs` module:
* `vector_by_offset`, `vectors_by_offsets` (was `pub(super)` on Segment)
* `read_vectors` (was `pub(crate)` on Segment in the now-deleted
  `segment/vectors.rs`)
* `vector` and `available_vectors_size_in_bytes` (`ReadSegmentEntry`
  trait methods)

`Segment::vector` and `Segment::available_vectors_size_in_bytes`
collapse to `with_view` delegators. The `retrieve` trait method now
calls `view.read_vectors(...)` inside per-name `with_view` closures.

Removed entirely:
* `segment/vectors.rs` (its only function moved to the view)
* `Segment::vector_by_offset` and `Segment::vectors_by_offsets`
  in `segment_ops.rs`

Kept on Segment:
* `vector_names` — trivial `vector_data.keys().cloned().collect()`,
  no shared logic to factor out.
* `all_vectors` — invariant `NamedVectors<'_>` lifetime fights the
  view borrow, and the body is a 6-line iteration over `self.vector`
  (which is itself now a view delegator). ReadOnlySegment will write
  the same trivial loop.

Tests in `segment/tests/mod.rs` that exercised
`segment.vector_by_offset(...)` now go through
`segment.with_view(|v| v.vector_by_offset(...))`.
